requirements.txt
aiohappyeyeballs==2.6.1
aiohttp==3.12.15
aiosignal==1.4.0
annotated-types==0.7.0
anyio==4.9.0
asttokens==3.0.0
async-timeout==5.0.1
attrs==25.3.0
cachetools==5.5.2
certifi==2025.4.26
charset-normalizer==3.4.2
decorator==5.2.1
dotenv==0.9.9
et_xmlfile==2.0.0
exceptiongroup==1.3.0
executing==2.2.0
frozenlist==1.7.0
genai==2.1.0
google-auth==2.40.2
google-genai==1.18.0
h11==0.16.0
httpcore==1.0.9
httpx==0.28.1
idna==3.10
ipython==8.18.1
jedi==0.19.2
lxml==6.0.0
matplotlib-inline==0.1.7
multidict==6.6.3
numpy==2.0.2
openai==0.27.10
openpyxl==3.1.5
pandas==2.3.1
parso==0.8.4
pexpect==4.9.0
pillow==11.2.1
prompt_toolkit==3.0.51
propcache==0.3.2
ptyprocess==0.7.0
pure_eval==0.2.3
pyasn1==0.6.1
pyasn1_modules==0.4.2
pydantic==2.11.5
pydantic_core==2.33.2
Pygments==2.19.2
python-dateutil==2.9.0.post0
python-docx==1.2.0
python-dotenv==1.1.1
pytz==2025.2
regex==2025.7.34
requests==2.32.3
rsa==4.9.1
six==1.17.0
sniffio==1.3.1
stack-data==0.6.3
tabulate==0.9.0
tiktoken==0.3.3
tqdm==4.67.1
traitlets==5.14.3
typing-inspection==0.4.1
typing_extensions==4.14.0
tzdata==2025.2
urllib3==2.4.0
wcwidth==0.2.13
websockets==15.0.1
yarl==1.20.1
